requirements.txt
beautifulsoup4
bitsandbytes
datasets
huggingface_hub
matplotlib
nltk
numpy
pandas
peft
pytest
python_dateutil
PyYAML
scikit_learn
sentencepiece
setuptools
torch
transformers
yfinance
lxml_html_clean 
trafilatura
finnhub-python
openai
sentence-transformers
gpt-researcher @ git+https://github.com/assafelovic/gpt-researcher.git
PyPortfolioOpt
diskcache
statsmodels
pmdarima
gitpython
alpaca-py